Add controls to set the day of the month and the enable / disable alarms. Looks nasty. Rework ssi and cgi to split into .h and .c files. Add Pico-Green-Clock.h file to pass over variable set/clear routines into the main code space.
