+++ html_files/index.shtml
@@ -6,13 +6,21 @@
     <body> <h1>PicoW Webserver Tutorial</h1>
         <br>
         <h2>This bit is SSI:</h2>
-        <p>Voltage: <!--#volt--></p>
-        <p>Temp: <!--#temp--> C</p>
+        <p>Voltage: <!--#volt--> V</p>
+        <p>Temp: <!--#temp--> Deg C</p>
         <p>LED is: <!--#led--></p>
+        <p>Day of Month is: <!--#day--></p>
+        <p>Alarm status is: <!--#alarm--></p>
         <br>
         <h2>This bit is CGI:</h2>
         <a href="/led.cgi?led=1"><button>LED ON</button></a>
         <a href="/led.cgi?led=0"><button>LED OFF</button></a>
+        <a href="/led.cgi?dayofmonth=1"><button>Month Day 1</button></a>
+        <a href="/led.cgi?dayofmonth=2"><button>Month Day 2</button></a>
+        <a href="/led.cgi?salarm=1"><button>Set Alarm 1</button></a>
+        <a href="/led.cgi?calarm=1"><button>Clear Alarm 1</button></a>
+        <a href="/led.cgi?salarm=2"><button>Set Alarm 2</button></a>
+        <a href="/led.cgi?calarm=2"><button>Clear Alarm 2</button></a>
         <br>
         <br>
         <a href="/index.shtml">Refresh</a>
